concert.sinks -> concert.coroutines
Sinks is too restrictive, e.g. flat correction is a coroutine but wants
to forward the calculated data.
