+++ concert/tests/unit/test_datatransfers.py
@@ -3,7 +3,7 @@ import numpy as np
 from threading import Event
 from concert.quantities import q
 from concert.helpers import coroutine, multicast, inject
-from concert.sinks import generate_sinograms
+from concert.coroutines import generate_sinograms
 from concert.tests.base import suppressed_logging, ConcertTest
 from concert.devices.motors.dummy import Motor
